// frontend/src/app/models/store-group.model.ts
export interface StoreGroup {
   id: number;
   name: string;
   address: string;
   tel_number: string;
   avatar: string;
   is_unique: boolean;
   created_at: string;
   updated_at: string;
   store_representatives_name: string;
   store_representatives_surname: string;
   store_representatives_email: string;
   store_representatives_cell_no: string;

}